Removed transforms
The rationale here is to remove (possibly) premature abstraction. The
value of the reference paths is very clear and worthy of being
built-in.

If paths are part of a transforms extension mechanism, then transforms
cannot be decomposed from reading, because include files need to have
their reference paths satisfied locally, prior to being presented for
inclusion.

Removing reference paths from transforms by making them 'core' means
that the transforms feature can be decomposed. I have opted to make
reference paths a metadata declaration, rather than a tag literal,
because of a hunch that it is better to keep tag literals simple and
processed at read-time, whereas reference paths are more complex and may
be usefully read explicitly if they are metadata.

Furthermore, the only other example transform left is that of format. I
think a new tag literal of #join accomplishes much of the value of
format, making format redundant. Therefore I have also removed the whole
transforms concept, as it is now not paying its way.