Introduce concept of "macros" to aero
For now, this is an internal refactoring only.  It introduces 3 kinds of
macro:

* `#or` - Instead of evaluating all of it's list of arguments, it only
evaluates them until it gets a truthy. The same as the clojure macro.

* `#profile`, `#user`, etc. - The case-like macros will only evaluate
the correct branch as appropriate.  These are also implemented using a
consistent underlying function which means this commit will fix #67.

* `#ref` is expressable using this macro system, and is no longer a
static dependency analysis.  The static analysis was fragile.  The
special-casing for it made it harder to implement other `#ref`-like
things such as a `#relref`. This will fix #50.