Remove deferred documentation
These will continue to work for the forseeable future, but they do not
serve a purpose any more.

Deferreds were introduced to balance the fact that conditional readers
(like `#or`) evaluate all branches.
Deferreds unfortunately do not compose with other readers.
This results in writing readers which expected a deferred and
automatically unwrap it.

Now that macros are introduced, we can avoid deferreds altogether.