Define alpha api for macro tag literals
When moving the multi-methods from aero.core, this has triggered some
bug in self-hosted with how vars are resolved.  In some cases, the
macro-namespace versions are being used instead of the normal namespace,
to workaround this I've introduced some macros from cgrand's macrovich.

These weren't necessary before, so I think it can be removed when
removing these out of alpha.